requirements.txt
fastmcp>=2.0.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
sqlalchemy[asyncio]
asyncpg
python-dotenv
uvicorn[standard]
python-dateutil
flask
flask-login
